Move pages to their own module.
