+++ payment-element-test-pages/gradle.properties
@@ -1,0 +1,1 @@
+STRIPE_ANDROID_NAMESPACE=com.stripe.paymentelementtestpages
