+++ payment-element-test-pages/src/main/java/com/stripe/paymentelementtestpages/EditPage.kt
@@ -1,11 +1,11 @@
-package com.stripe.android.testing
+package com.stripe.paymentelementtestpages
 
 import androidx.compose.ui.test.SemanticsNodeInteraction
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsEnabled
 import androidx.compose.ui.test.assertIsNotEnabled
 import androidx.compose.ui.test.hasTestTag
-import androidx.compose.ui.test.junit4.AndroidComposeTestRule
+import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import com.stripe.android.paymentsheet.ui.REMOVE_BUTTON_LOADING
@@ -18,7 +18,7 @@ import com.stripe.android.uicore.elements.DROPDOWN_MENU_CLICKABLE_TEST_TAG
 import com.stripe.android.uicore.elements.TEST_TAG_DROP_DOWN_CHOICE
 
 class EditPage(
-    private val composeTestRule: AndroidComposeTestRule<*, *>
+    private val composeTestRule: ComposeTestRule
 ) {
     fun waitUntilVisible() {
         composeTestRule.waitUntil {
@@ -90,7 +90,13 @@ class EditPage(
         if (waitUntilComplete) {
             composeTestRule.waitUntil(timeoutMillis = 5_000L) {
                 composeTestRule
-                    .onAllNodes(hasTestTag(UPDATE_PM_SAVE_BUTTON_TEST_TAG).and(hasTestMetadata("isLoading=true")))
+                    .onAllNodes(
+                        hasTestTag(UPDATE_PM_SAVE_BUTTON_TEST_TAG).and(
+                            hasTestMetadata(
+                                "isLoading=true"
+                            )
+                        )
+                    )
                     .fetchSemanticsNodes()
                     .isEmpty()
             }
