+++ payment-element-test-pages/src/main/java/com/stripe/paymentelementtestpages/ManagePage.kt
@@ -1,9 +1,9 @@
-package com.stripe.android.testing
+package com.stripe.paymentelementtestpages
 
 import androidx.compose.ui.test.hasAnyDescendant
 import androidx.compose.ui.test.hasTestTag
 import androidx.compose.ui.test.isSelected
-import androidx.compose.ui.test.junit4.AndroidComposeTestRule
+import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
@@ -15,7 +15,7 @@ import com.stripe.android.paymentsheet.verticalmode.TEST_TAG_MANAGE_SCREEN_SAVED
 import com.stripe.android.paymentsheet.verticalmode.TEST_TAG_SAVED_PAYMENT_METHOD_ROW_BUTTON
 
 class ManagePage(
-    private val composeTestRule: AndroidComposeTestRule<*, *>,
+    private val composeTestRule: ComposeTestRule,
 ) {
     fun waitUntilVisible() {
         composeTestRule.waitUntil {
@@ -80,7 +80,15 @@ class ManagePage(
     fun assertCardIsVisible(paymentMethodId: String, cardBrand: String) {
         composeTestRule.onNode(
             hasTestTag("${TEST_TAG_SAVED_PAYMENT_METHOD_ROW_BUTTON}_$paymentMethodId")
-                .and(hasAnyDescendant(hasTestTag(TEST_TAG_ICON_FROM_RES).and(hasTestMetadata(cardBrand)))),
+                .and(
+                    hasAnyDescendant(
+                        hasTestTag(TEST_TAG_ICON_FROM_RES).and(
+                            hasTestMetadata(
+                                cardBrand
+                            )
+                        )
+                    )
+                ),
             useUnmergedTree = true,
         ).assertExists()
     }
