+++ payment-element-test-pages/src/main/java/com/stripe/paymentelementtestpages/TestMetadata.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.testing
+package com.stripe.paymentelementtestpages
 
 import androidx.compose.ui.test.SemanticsMatcher
 import com.stripe.android.paymentsheet.utils.TestMetadata
