+++ paymentsheet/build.gradle
@@ -58,6 +58,7 @@ dependencies {
     testImplementation project(':financial-connections')
     testImplementation project(':network-testing')
     testImplementation project(':payments-core-testing')
+    testImplementation project(':payment-element-test-pages')
     testImplementation project(':screenshot-testing')
 
     testImplementation testLibs.androidx.archCore
@@ -89,6 +90,7 @@ dependencies {
 
     androidTestImplementation project(':network-testing')
     androidTestImplementation project(':payments-core-testing')
+    androidTestImplementation project(':payment-element-test-pages')
 
     androidTestImplementation libs.places
     androidTestImplementation testLibs.androidx.composeUi
