+++ settings.gradle
@@ -6,6 +6,7 @@ include ':identity'
 include ':identity-example'
 include ':hcaptcha'
 include ':network-testing'
+include ':payment-element-test-pages'
 include ':payment-method-messaging'
 include ':payments'
 include ':payments-core'
